packages/txtminer/pyproject.toml
[project]
name = "txtminer"
version = "0.1.0"
description = "Extract service from domain validation TXT record"
readme = "README.md"
authors = [
    { name = "Julien OPPLIGER", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "common",
    "dnspython>=2.8.0",
    "open-asset-model>=1.1.1",
    "open-asset-store>=0.0.2",
    "termcolor>=3.3.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
common = { workspace = true }

[project.scripts]
txtminer = "txtminer.__main__:main"
